refactor(foldkit): relocate Subscription, ManagedResource, patchVNode
These lived under runtime/ but are not runtime machinery. Each was only
parked there because the runtime consumes it, while its own module reached
back across the folder boundary to re-export it.

- Subscription: the make/aggregate/lift/persistent constructors and their
  types move from runtime/subscription.ts to subscription/subscription.ts.
- ManagedResource: the make/lift/aggregate constructors move from
  runtime/managedResource.ts and merge with the tag/ResourceNotAvailable
  core, so the whole module now lives in managedResource/managedResource.ts
  behind a barrel index.ts.
- patchVNode: a thin wrapper over vdom primitives (dedupe + snabbdom patch)
  moves from runtime.ts to vdom.ts alongside the functions it calls.

No public API change: every package subpath still resolves through the same
public.ts, exporting the same names. The ManagedResource constructor unit
test moves with the code; the runtime integration tests stay in runtime/.
