+++ .changeset/relocate-runtime-parked-modules.md
@@ -1,0 +1,4 @@
+---
+---
+
+Internal refactor with no public API change: relocate the Subscription and ManagedResource modules and the patchVNode helper out of runtime/ into their own modules. No release needed.
