+++ packages/foldkit/src/managedResource/public.ts
@@ -1,17 +1,18 @@
-export { tag, ResourceNotAvailable } from './index.js'
+export {
+  tag,
+  ResourceNotAvailable,
+  make,
+  lift,
+  aggregate,
+} from './managedResource.js'
 
 export type {
   ManagedResource,
   ManagedResourceService,
   Value,
   ServiceOf,
-} from './index.js'
-
-export { make, lift, aggregate } from '../runtime/managedResource.js'
-
-export type {
   Entry,
   ManagedResources,
   ManagedResourceConfig,
   ServicesOf,
-} from '../runtime/managedResource.js'
+} from './managedResource.js'
